+++ pycrud/crud/_core_crud.py
@@ -4,18 +4,18 @@ from typing import Type, Iterable, List
 
 from pycrud.crud.query_result_row import QueryResultRow, QueryResultRowList
 from pycrud.query import QueryInfo
-from pycrud.types import RecordMapping, IDList
-from pycrud.values import ValuesToWrite
+from pycrud.types import Entity, IDList
+from pycrud.values import ValuesToUpdate
 
 
 @dataclass
 class CoreCrud:
     @abstractmethod
-    async def insert_many(self, table: Type[RecordMapping], values_list: Iterable[ValuesToWrite], *, _perm=None) -> IDList:
+    async def insert_many(self, table: Type[Entity], values_list: Iterable[ValuesToUpdate], *, _perm=None) -> IDList:
         pass
 
     @abstractmethod
-    async def update(self, info: QueryInfo, values: ValuesToWrite, *, _perm=None) -> IDList:
+    async def update(self, info: QueryInfo, values: ValuesToUpdate, *, _perm=None) -> IDList:
         pass
 
     @abstractmethod