+++ pycrud/crud/query_result_row.py
@@ -5,7 +5,7 @@ from pycrud.utils.name_helper import get_class_full_name
 
 if TYPE_CHECKING:
     from pycrud.query import QueryInfo
-    from pycrud.types import RecordMapping
+    from pycrud.types import Entity
 
 
 @dataclass
@@ -14,7 +14,7 @@ class QueryResultRow:
     raw_data: Union[Tuple, List]
     info: 'QueryInfo'
 
-    base: Type['RecordMapping']
+    base: Type['Entity']
     extra: Any = field(default_factory=lambda: {})
 
     def __post_init__(self):
@@ -24,7 +24,7 @@ class QueryResultRow:
         if self._dict_cache is None:
             data = {}
             for i, j in zip(self.info.select_for_crud, self.raw_data):
-                if i.table == self.base:
+                if i.entity == self.base:
                     data[i.name] = j
 
             if self.extra:
@@ -42,10 +42,10 @@ class QueryResultRow:
         return self._dict_cache
 
     def __repr__(self):
-        return '<%s %s id: %s>' % (self.__class__.__name__, get_class_full_name(self.info.from_table), self.id)
+        return '<%s %s id: %s>' % (self.__class__.__name__, get_class_full_name(self.info.table), self.id)
 
 
-class QueryResultRowList(list):
+class QueryResultRowList(List[QueryResultRow]):
     def __init__(self, *args):
         super().__init__(*args)
         self.rows_count = None