+++ tests/test_query.py
@@ -4,10 +4,10 @@ from typing import Any, Union, List
 
 from pycrud.pydantic_ext.hex_string import HexString
 from pycrud.query import QueryInfo
-from pycrud.types import RecordMapping, RecordMappingField
+from pycrud.types import Entity, EntityField
 
 
-class User(RecordMapping):
+class User(Entity):
     id: int
     nickname: str
     username: str
@@ -16,7 +16,7 @@ class User(RecordMapping):
 
 
 def test_query_select_exclude():
-    q = QueryInfo.from_table_raw(User, select=[User.id, User.nickname, User.password], select_exclude=[User.nickname])
+    q = QueryInfo.from_table(User, select=[User.id, User.nickname, User.password], select_exclude=[User.nickname])
     assert q.select_for_crud == [User.id, User.password]