+++ tests/test_values.py
@@ -2,11 +2,11 @@ from typing import Optional, List
 
 from pydantic import Field
 
-from pycrud.types import RecordMapping
-from pycrud.values import ValuesToWrite, ValuesDataFlag
+from pycrud.types import Entity
+from pycrud.values import ValuesToUpdate, ValuesDataUpdateFlag, ValuesToCreate
 
 
-class User(RecordMapping):
+class User(Entity):
     id: Optional[int]
     nickname: str
     test: int = 111
@@ -14,23 +14,24 @@ class User(RecordMapping):
 
 
 def test_values_bug_1():
-    v = ValuesToWrite({
+    v = ValuesToUpdate({
         'nickname': 'aaa'
     })
-    v.bind(False, User)
+    v.bind(User)
     assert 'test' not in v
 
 
 def test_values_array_extend():
-    v = ValuesToWrite({
+    v = ValuesToUpdate({
         'arr.array_extend': ['aa', 'bb']
-    }, table=User, try_parse=True)
-    assert v.data_flag['arr'] == ValuesDataFlag.ARRAY_EXTEND
+    }, entity=User).bind()
+    assert v.data_flag['arr'] == ValuesDataUpdateFlag.ARRAY_EXTEND
 
 
 def test_values_try_bind():
-    v = ValuesToWrite({
+    # TODO: try bind 不存在了 但之前为什么会有这个阶段？
+    v = ValuesToUpdate({
         'aaa': 'bbb',
         'arr.array_extend': ['aa', 'bb']
-    }, table=User, try_parse=True)
-    assert v['aaa'] == 'bbb'
+    }, entity=User).bind()
+    assert 'aaa' not in v